requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Flask-Cors==4.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
